backend/requirements.txt
# FastAPI Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
pymysql==1.1.0
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
pydantic-settings==2.1.0

# AI & ML
openai==1.3.7
groq==0.4.1

# HTTP Requests
httpx==0.25.2
requests==2.31.0

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Utilities
pydantic==2.5.2
typing-extensions==4.8.0
gunicorn==21.2.0
